renamed tokenMiddleware for consistency
