+++ locksmith/src/app.js
@@ -2,7 +2,7 @@ const bodyParser = require('body-parser')
 const cors = require('cors')
 const express = require('express')
 const Sequelize = require('sequelize')
-const tokenMiddleware = require('./token_middleware')
+const tokenMiddleware = require('./tokenMiddleware')
 const Lock = require('./lock')
 const logger = require('./locksmithLogger')
