+++ locksmith/src/tokenMiddleware.js
